Pr fixes
